% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census.R
\docType{data}
\name{census}
\alias{census}
\title{census}
\format{
This contains 61 measurements with 5 variables
\describe{
  \item{total_population}{thousands}
  \item{professional_degree}{percent}
  \item{employed_age_over16}{percent}
  \item{government_employment}{percent}
  \item{median_home_value}{one hundred thousand dollars}
  ...
}
}
\source{
Applied Multivariate Statistical Analysis tth edition
}
\usage{
census
}
\description{
This dataset provided information, by tract, on five socioeconomic variables for the Madison, Wisconsin, area.
This contains 61 measurements with 5 variables
}
\keyword{datasets}
