% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcaAnalysis.R
\name{pcaAnalysis}
\alias{pcaAnalysis}
\title{pcaAnalysis}
\usage{
pcaAnalysis(data)
}
\arguments{
\item{data}{the data we try to find the principle component}
}
\value{
the list
}
\description{
pcaAnalysis
}
\examples{
pcaAnalysis(census)
}
